feat: Add E2E Local Node Smart Contracts Workflow (#842)
This PR adds a workflow that runs end to end tests using a local Hedera node setup. The workflow:

-  Builds our block-node Docker image from the current commit, to validate that there are no regressions.
-  Sets up the local-node environment with our built image.
-  Runs selected Hedera Smart Contracts tests and JSON-RPC Relay Acceptance tests. We are using only selected for two reasons:
   a. To lower running time
   b. We try to test and run transaction of different types. Most of the tests are running the same transactions.
-  Checks block-node logs for errors

The workflow runs on push to main/release branches and can be triggered manually. Also has option to select custom tags for CN, MN and Relay. Also to select custom branch to test for regressions.

Signed-off-by: georgi-l95 <glazarov95@gmail.com>