Updating Design Doc for Backfill, updating Metrics doc and Configuration doc.
While updating configuration, realized that the current file was very much outdated since it was still referencing pre re-do of BN, so I updated to match current state.

Also, simplified VerificationConfig and leave it empty since we know in the future it will be needed so we might as well keep the underlying plumbing.

Signed-off-by: Alfredo Gutierrez Grajeda <alfredo@hashgraph.com>